InventoryManagementSystem/VarietyStore/requirements.txt
<<<<<<< HEAD
asgiref==3.8.1
barcode==1.0.4
Django==5.1.6
django-bootstrap5==24.3
django-widget-tweaks==1.5.0
Levenshtein==0.26.1
pillow==11.1.0
psycopg2==2.9.10
python-barcode==0.15.1
python-Levenshtein==0.26.1
RapidFuzz==3.12.1
setuptools==75.8.0
sqlparse==0.5.3
tzdata==2025.1
=======
asgiref==3.8.1
barcode==1.0.4
Django==5.1.6
django-bootstrap5==24.3
django-widget-tweaks==1.5.0
Levenshtein==0.26.1
pillow==11.1.0
psycopg2==2.9.10
python-barcode==0.15.1
python-Levenshtein==0.26.1
RapidFuzz==3.12.1
setuptools==75.8.0
sqlparse==0.5.3
tzdata==2025.1
>>>>>>> c7a1200 (Changes)
